+++ test/libsolidity/syntaxTests/constructor/constructible_internal_constructor.sol
@@ -1,6 +1,8 @@
 abstract contract C {
-	constructor() {}
+	constructor() internal {}
 }
 contract D is C {
 	constructor() { }
 }
+// ----
+// Warning 2462: (23-48): Visibility for constructor is ignored. If you want the contract to be non-deployable, making it "abstract" is sufficient.