+++ test/libsolidity/syntaxTests/constructor/external_constructor.sol
@@ -2,4 +2,4 @@ contract test {
 	constructor() external {}
 }
 // ----
-// SyntaxError 2462: (17-42): Visibility specified for constructor. If you want the contract to be non-deployable, make it "abstract".
+// TypeError 9239: (17-42): Constructor cannot have visibility.