+++ test/libsolidity/syntaxTests/constructor/inconstructible_abstract_contract.sol
@@ -1,0 +1,8 @@
+abstract contract C {
+	constructor() {}
+}
+contract D {
+	function f() public { C c = new C(); c; }
+}
+// ----
+// TypeError 4614: (84-89): Cannot instantiate an abstract contract.